Avoid NoMethodError exceptions in delegating Response methods
I noticed that exceptions like this one [1] were occurring on staging.

I discovered that this is possible if the attachment data belongs to a
consultation response which in turn belgongs to a deleted consultation
which in turn has a previous edition. I've added this scenario to
AttachmentDataVisiblityTest and called AttachmentData#deleted? to
trigger the exception.

In fixing the problem, I noticed a bunch more methods on Response which
I've added recently are susceptible to the same problem. I've driven out
fixes for them by adding tests to ResponseTest rather than making
AttachmentDataVisiblityTest even longer.

I've reused AttachmentData::NullAttachable to avoid duplicating the
default return values.

Note that Response#access_limited_object is not vulnerable to the same
problem, because unlike the others, it doesn't call a method on the
parent consultation. The same is true of all the equivalent methods on
PolicyGroup.

[1]: https://sentry.io/govuk/app-whitehall/issues/491627163/