Call AssetManagerUpdateAssetWorker synchronously
We are now calling AssetManagerAttachmentAccessLimitedWorker
asynchronously and there's no need to create additional
AssetManagerUpdateAssetWorker jobs from this worker. In fact we want
to avoid doing so as there may be a change in the data between when
the job is queued and when it is executed.

AssetManagerUpdateAssetWorker is used elsewhere, so rather than inline
the logic we call it synchronously using `perform`. Note `perform`
unlike `perform_async` does not convert symbols to strings so we have
to make that change at the call site.