Remove redundant Sidekiq::Testing.fake! blocks from tests
In #3793 Sidekiq's testing mode was set to "fake" by default, so these
blocks are unnecessary.

The diff is best viewed using the `--ignore-all-space` option, because
of all the indentation changes.

Note that I couldn't remove the Sidekiq::Testing.fake! blocks from
RequestTracingTest, because the whole test is being run in "inline"
mode.