Assert setup complete in AttachmentAccessLimitedIntegrationTest
When setting up the state of the app via the user interface, I think
it's sensible to add some kind of assertion that this has completed
successfully before moving on to make the actual assertions. Although
this couples the tests to messages displayed in the user interface, I
think the increased reliability is worthwhile. And if the duplication is
really a concern it could be substantially reduced by refactoring.

Note that the change to the `add_file_attachment` method in the test is
needed in the "when attachment is replaced" scenario, because the
original attachment title is used in the flash notice which I want to
assert against.