Introduce AssetManagerAttachmentMetadataUpdater
The `update_range` method in this class calls the various asset
manager attachment update workers with every attachment data within a
range of ids. We plan to call this from a rake task to update the
metadata of all the attachments that have been uploaded to Asset
Manager. We need to do this before we can switch over to serving them
from Asset Manager instead of Whitehall.