Fix for exception in PublishingApi::PayloadBuilder::AccessLimitation.for
The new integration test scenario where an attachment is added to an
outcome on an access-limited consultation triggered a NoMethodError
exception as described in this comment [1]. As mentioned there, this
root cause of this is similar to an exception that's been seen in
production [2].

This is a naive fix which allows
PublishingApi::PayloadBuilder::AccessLimitation.for to obtain the
organisations for a (consultation) Response by delegating to its parent
Consultation.

I'm really not very happy with the overall implementation of this
access-limiting behaviour; I feel as if the logic is rather scattered
around the codebase. However, I'll leave that for another commit or PR.

Note that a number of AssetManagerCreateWhitehallAssetWorker jobs have
have failed in production, which means that assets will not have been
created in Asset Manager for the relevant attachments. However, assuming
the Sidekiq workers are restarted after a deployment, if any of them are
still on the retry queue, they should succeed after this change has been
deployed. Although there's a slight question in my mind about how long
the files will have hung around in the Whitehall.asset_manager_tmp_dir
directory.

[1]:
https://github.com/alphagov/asset-manager/issues/522#issuecomment-373049045
[2]: https://sentry.io/govuk/app-whitehall/issues/490263226/