Fix for AttachmentData#significant_attachment bug
In order to fix this issue [1], we should be ignoring attachments with
no attachable, i.e. attachments associated with deleted editions.

We've added a realistic/representative scenario to
AttachmentDataVisibilityTest and more thorough & more unit-y tests in
AttachmentDataTest to drive out the changes.

We compared the difference between `AttachmentVisibility#visible?` and
`AttachmentData#visible_to?` for all `AttachmentData` records in
integration (about 500,000). There were a number of differences before
this change, and no differences after so we're confident we're mirroring
the previous behaviour provided by `AttachmentVisibility`.

[1]: https://github.com/alphagov/asset-manager/issues/524