Convert if/else to guard condition in attachment controller actions
I think this makes the "happy" path more explicit and brings this
controller actions more into line with the media controller download
actions in Asset Manager.

I believe there is some duplication in the logic for the non-happy paths
and I hope to reduce this in subsequent commits. I think this change
will make that easier.