Respond with 404 Not Found for infected image attachment
Previously a request for an image attachment was redirected to the
placeholder image if it was unscanned *or* infected. This was
inconsistent with the behaviour for a non-image attachment which is only
redirected to a placeholder URL if the asset is unscanned. This commit
brings the behaviour for image attachments into line with that for
non-image attachments.

Note that this is also consistent with the behaviour in the Asset
Manager media controller download actions where a request for an
infected asset always results in a 404 Not Found.