Use guard condition to check for CSV file in CsvPreviewController#show
There are a number of scenarios in which the response is a 404 Not
Found. I'm hoping to extract the logic for each of these scenarios into
separate guard conditions to make things more explicit and to more
closely match the logic in the Asset Manager media controller download
actions. This is a first step in that direction.

Note that there is a subtle change in behaviour here, because previously
if a non-CSV attachment had also been unpublished, replaced, or
unscanned, then one of the redirect conditions would've kicked in first.
However, I don't think this is a big deal and we can always re-order the
guard clauses later.