Fix bug in AttachmentData#visible_to? for replaced attachments
This bug meant that unless there was some reason why an AttachmentData
was considered not "visible", requests for a replaced attachment
would've resulted in the original attachment being served.

Thus a replaced attachment on a draft edition was redirected correctly,
because being draft meant that it was not visible. But a replaced
attachment on a published edition was not redirected, because being
published meant that it was visible.

I believe this has been the case for a long time, because
AttachmentVisiblity#visible? did not take into account whether an
AttachmentData had been replaced or not.

Both AttachchmentsController#show & CsvPreviewController#show use
BaseAttachmentsController#attachment_visible? and thence
BaseAttachmentsController#upload_exists? and AttachmentData#visible_to?
to decide whether to render the attachment/preview or to call
BaseAttachmentsController#fail. It is the latter which contains the
condition to redirect requests for replaced attachments.

Unfortunately the controller tests for replaced attachments were only
passing by accident, because BaseAttachmentsController#upload_exists?
and/or AttachmentData#visible_to? were returning false for a number of
reasons; whereas BaseAttachmentsController#upload_exists? should've been
returning true and AttachmentData#visible_to? should've only been
returning false, because AttachmentData#replaced? was returning true.

These tests were also faulty when AttachmentVisibility was in place,
for similar reasons. BaseAttachmentsController#upload_exists? was again
returning false, because the attachment file has not been virus-scanned
and AttachmentVisibility#visible? was only returning false, because the
AttachmentData was an orphan and had no parent edition.

I've updated the controller tests so that they're more realistic and
they forced me to add the dependency on AttachmentData#replaced? in the
implementation of AttachmentData#visible_to?. I'm not completely happy
with the amount of stubbing I've had to do, but I'm hoping I can clear
it up later.