Ignore missing files when updating metadata in Asset Manager
When testing `AssetManagerAttachmentMetadataUpdater.update_range` in
integration we ran into some problems where `AttachmentData` objects
didn't have a corresponding file on the file system. The lack of file
meant that the asset hadn't been created in Asset Manager (because we
populated Asset Manager using the files on disk) and so we saw
exceptions when trying to retrieve the asset from Asset Manager (in
`AssetManagerWorkerHelper#find_asset_by`).

Some of these attachments have been replaced which means that, despite
not having a file on disk, requests for them are redirected to their
replacement. We've identified these attachments and are creating assets
for them in Asset Manager as part of
https://github.com/alphagov/asset-manager/issues/525. Once these
records have been created we're confident that it's safe to ignore any
remaining `AttachmentData` records that don't have files.