Re-order checks in AttachmentsController
I've added a bunch of tests which constrain the implementation of
AttachmentsController#show to work the same way as it did before this PR
in terms of the order in which the conditions are evaluated.

There weren't really any such tests before I did the re-write of
AttachmentsControllerTest at the beginning of this PR and in hindsight I
wish I'd added them at that point. However, that would mean re-working a
bunch of the commits which refactor the controller and so I'm adding
them after the fact.

Note that I'm not convinced that this order is entirely sensible, but
I'm pretty confident it matches the original behaviour.