Call update workers synchronously
The problem with calling these workers asynchronously is that it is
possible that AssetManagerAttachmentDeleteWorker gets executed first
and, if required, soft-deletes the asset in Asset Manager. This means
that if any of the other jobs queued with this attachment_data ID are
subsequently executed the call to Asset Manager to GET the asset will
return 404.

Although a deleted asset cannot be served to the user, it is possible
to "undelete" assets in Asset Manager. This means we want to ensure
that any additional metadata is set before the asset is deleted.

Changing this worker to call the other works synchronously means that
we can control the execution order. We'll change the order in the next commit.