Revert "Merge pull request #3881 from alphagov/refactor-attachment-controllers"
This reverts commit 386ad20bfe00c87b2f8e76295d457ec305cab2b1, reversing
changes made to 3849ca03992d33b83738725c9ad91e52c799954d.

These changes were based on a faulty assumption, i.e. that
AttachmentVisibility#visible? always returned false if the
AttachmentData had been replaced. This means these changes will get in
the way of us fixing a problem in master where requests for attachments
replaced by an attachment on a draft edition are redirecting to the
replacement which then responds with a 404 Not Found; whereas it should
serve the original attachment.

We think this problem was introduced in #3874.