Add scenarios to AttachmentReplacementIntegrationTest
This would've caught the problem that was accidentally introduced in
this PR [1] and reverted in this PR [2]. More specifically the assertion
on line 92 would've failed, because the response would've incorrectly
been a 301 Moved Permanently redirect.

It also serves to illustrate how the attachment replacement
functionality will work once we switch over to serving Whitehall
attachments from Asset Manager. In particular it highlights how we will
be dependent on the functionality introduced into Asset Manager in this
PR [3], i.e. Whitehall will mark the relevant Asset Manager assets as
replaced at the time where the attachment is replaced, even though in
some circumstances the redirect to the replacement should not happen.

[1]: https://github.com/alphagov/whitehall/pull/3874
[2]: https://github.com/alphagov/whitehall/pull/3894
[3]: https://github.com/alphagov/asset-manager/pull/531