Inline CsvFileFromPublicHost#response
This private method is only used in the constructor and so inlining it
into a local variable shouldn't cause any problems and I'm hoping it
will make it easier to do some refactoring that I'm planning.

Note that previously we were always requesting the first bit of the CSV
file twice, because we were calling the #response method twice and it's
result was not memo-ized.