Avoid dependency on attachment path for tempfile prefix
It's not obvious to me why we need to generate the tempfile prefix from
the attachment path. The tempfile suffix should be good enough to
guarantee uniqueness.

Note that the truncation of the path was added in 18bc4340b7c to fix a
bug with tempfiles generated from attachments with very long file
names. This change ensures we won't have that problem in the future,
and we're confident that the test added to catch any regression can be
removed.