Remove dependency on filesystem from CsvPreviewController#show
We now use the response from the range request for the CSV file to
determine whether it is clean, infected, unscanned, or missing.

Removing the underlying file in the setup block demonstrates that we are
no longer dependent on the file being present.

This should make it easier to switch over to use the AssetManagerStorage
in the AttachmentUploader.