Expand show/hide taxonomy UI to other models
Statistical Data Sets and Fatality Notices support related policies but
don't expose them in the UI. Some conditional logic was necessary to
support this in the generic helper.

Fatality Notices can only be created with specific users/orgs, which is
why the user is now built as a GDS Editor, which has permission to
create Fatality Notices (and other doc types).