Redirect attachment requests to asset host
It's currently possible to request Whitehall attachments at
/government/uploads/attachment_data/file via both www.gov.uk and the
asset host (e.g.  assets.publishing.gov.uk). We want all asset requests
to go via the asset host so that we can make changes to the config in a
single place.

I'm using `Plek#public_asset_host` as it returns "The public-facing
asset base URL"[1]. I've tested this on integration and confirmed that
it returns
'https://assets-origin.integration.publishing.service.gov.uk'.

This is very similar to the redirect added in
0aa298afeb19a55dc56954712729af585ee2568c.

[1]: https://github.com/alphagov/plek/blob/4d63a574ab36e9e8833ff14471482d1cbb110e61/lib/plek.rb#L97