Make incoming_upload_exists? check the path only
The host of the Location header for an unscanned CSV attachment can
depend on the publication status of the asset. For example when
testing a newly uploaded CSV in integration the Location header was
`https://www-origin.integration.publishing.service.gov.uk/government/placeholder`
while the placeholder URL was
`https://whitehall-admin.integration.publishing.service.gov.uk/government/placeholder`. Comparing
the paths to determine whether the CSV is in the process of being
virus scanned is good enough and avoids adding any additional logic.