Avoid exceptions if asset is deleted in Asset Manager & Whitehall
If we can't find the asset in Asset Manager (i.e. we get a
`GdsApi::HTTPNotFound` exception) then it means that the asset has been
deleted or that it never existed. In either case, if the asset has also
been deleted from Whitehall then there's nothing to do as a request for
the asset in either app will result in a 404 so the behaviour remains
the same.

Note that it's theoretically possible for Whitehall to have marked an
asset as deleted in Asset Manager before then trying to update some
other metadata about it (e.g. its access-limited state). If this
happens, and if we were then to undelete the asset in Asset Manager; it
might not be in the state we'd expect. We're not worried about this at
the moment as there's currently no way through the user-interface to
undelete assets from Whitehall.