Avoid setting draft to true on replaced assets in Asset Manager
Currently we are seeing exceptions from the
AssetManagerAttachmentDraftStatusUpdateWorker when we attempt to set
`draft` to `true` on an asset that has already been replaced. This
raises an GdsApi::HTTPUnprocessableEntity exception with the
message (from Asset Manager) "Draft cannot be true, because already
replaced".

Assets in Whitehall that have public URLs that redirect to other
assets can be "draft" in Whitehall as the code doesn't enforce this
edge case. In Asset Manager however we are a bit stricter and raise
the above exception to notify the clients.

This commit avoids raising the exception in the first place by setting
the 'draft' flag to `false` in Asset Manager in this case.