Change legacy tags worker to fetch topic taxons
Previously the taxon_ids to work on were passed as part of the arguments
to the job, which could result in a race condition if further updates to
the same document occur in the near future. Fetching the taxons for the
document at the point of execution helps ensure they are up to date.