Change legacy tags worker to suppor patch versions
Now that we are retrieving the links for the edition inside the worker,
we have the version of the content as part of the response. This can be
passed to the subsequent patch_links call to ensure the update is a
transaction (the content was not changed between fetching the existing
links and updating the new ones).