Add rake task to regenerate missing thumbnails.
When running the AssetManagerAttachmentMetadataUpdater task we
encountered a number of exceptions that were due to thumbnail files
not being found in Asset Manager. After investigation we noticed
that these files are not present on the NFS mount for Whitehall.

Our migration of all files from Whitehall to Asset Manager was based
on the files on disk so this explains why these are missing. Our best
guess is that something either went wrong with the thumbnail
generation[1] or these assets were added before the thumbnail
generation code was made more robust. In either case it seems simplest
to simply regenerate the missing thumbnails and allow them to be
uploaded to Asset Manager. The rake task in this commit does that for
the attachment ids extracted from the exceptions in Sentry[2] that
were due to missing thumbnails.

[1]
https://github.com/alphagov/whitehall/blob/0545394231b8a028935116e1ca8740b96591af1c/app/uploaders/attachment_uploader.rb#L55
[2] https://sentry.io/govuk/app-whitehall/issues/504366848/