Avoid redirecting to asset host for signed in users
The redirection added in PR 3918 didn't take draft attachments into
account; in that draft attachments aren't accessible via the asset host.
This means that draft attachments (e.g. attachments added to a draft
consultation) aren't visible from within whitehall-admin.

Rather than check the state of the attachment I've decided to avoid
redirecting if the user is signed in. This means that if you're signed
in you'll continue to view all attachments on whitehall-admin or
draft-origin as you would have done previously.

I've had to update the "redirects asset requests that aren't made via
the asset host when the filename contains multiple periods" integration
test by logging out as we login as an admin user in the test setup.