Handle soft deleted assets in Asset Manager
The Asset Manager API was recently updated so that the API responds to
GET requests for deleted assets with a 200 instead of a 404 (it includes
the deleted state in the JSON response). This change to the
`AssetManagerUpdateAssetWorker` treats assets that have been deleted in
Asset Manager the same as assets that don't exist.