Avoid rescuing GdsApi::HTTPNotFound exception
Rescuing `GdsApi::HTTPNotFound` was a proxy for handling assets in Asset
Manager that had been deleted. Now that we can identify these deleted
assets we no longer need to rescue these exceptions: we expect these
assets to exist in Asset Manager so we want to be notified if they can't
be found.