Initialize Mocha before stubbing in Cucumber Before blocks
Since Mocha v1.5.0, it's not possible to use Mocha until it has been
explicitly initialized by a call to Mocha::Hooks#mocha_setup. Currently
this is being called from a Before block in `features/support/mocha.rb`.

In order to avoid many of the features failing, we need to ensure that
this Before block is run *before* any stubbing is done in these other
Before blocks. To achieve that I've explicitly required them in the
appropriate files in `features/support`. This fixes the failing
features.