Only send international delegations to Rummager
We want Rummager to index the new world location taxons so we need
to stop Whitehall from sending them to Rummager as they would show
as duplicates in the search results. This is because Whitehall
sends documents to the 'government' index while Rummager adds the
taxons to the 'govuk' index.