Send people to the publishing-api
This makes sure that we're sending people to the publishing-api with
the `person` schema. The schema is a bit bare bones at the moment so
we're only sending the persons image. The rest can be added later.

Needs https://github.com/alphagov/govuk-content-schemas/pull/749 to be
deployed before the tests pass

https://trello.com/c/xebVLRcp