Handle scenario when there are no reports to send
In the unlikely event that Whitehall had no bad links to report on this
script was erroring by trying to zip files that didn't exist.

I decided to put this in a conditional to stop the erroring.

An alternative approach would be to send out empty CSV files (which
could be a preferred approach if users anticipate emails) however this
seemed like a larger change in behaviour so I decided against it.