Separate out the World taxons from the Topic Taxonomy
In the Taxonomy::PublishingApiAdapter. This allows removing the World
taxons from the Topic Taxonomy, while retaining the ability to tag to
them in Whitehall.

Having the World taxons in the Topic Taxonomy is unusual, as they
don't categorise content by what it is about, but rather by
country. Removing them makes the Topic Taxonomy better, and will help
improve the metrics that we have for it, as they will no longer
contain the World taxon.