Allow a null batch_id on link_checker_api_reports
This is to allow us to create LinkCheckerApiReports that don't go to
Link Checker API as the edition has no links. Whilst this may seem like
a strange situation it is implemented to cut back on all the logic as to
whether a link check report can be used.

Currently if an edition has links and then has them removed it could
have an out of date LinkCheckerApiReport associated with it that could
be revealed in some places and not others.