Alter publishing flow
Instead of creating a new document and then being taken to the overview
edit page. The user will now be redirected to the topic tagging page if
the documents is eligible to be tagged to the topic taxons. If the
document is not eligible, the user will be redirected to the legacy
tagging page. After saving the tags, the publisher will be taken to the
overview edit page.