Separate world from topic taxons in PublishingApiAdapter
This class gets data from Publishing Api about the topic taxons available
in the publishing-api. This data is stored in a Redis cache, and it's this
cache that's used by Whitehall to determine which topic taxons can be tagged
to.

This change will make it easier to create a new cache of world taxons in Redis
in the next commit.