Create a new class EditionWorldTaxonLinkPatcher
This started as duplicate of EditionTaxonLinkPatcher. This new
class gets any world taxons that are already tagged to the
edition. This class will be use in a later commit that creates
a new controller for displaying and tagging to the World taxonomy.

It will be used to populate the lists of taxon branches that can be
tagged to.