Create a form to tag to the World taxonomy
This form is where users will tag to the world taxonomy and submit their changes
The controller to route them to this page, and process the results of the tagging
will be created in the next commit.

This is a duplicate of TaxonomyTagForm. It was not possible to re-use that
class as it calls on TopicTaxonomy to embed "invisible taxons" in a hidden
field in the view.

Right now there aren't any hidden world taxons, so this field isn't needed
and it didn't feel right to populate the World taxonomy form with data from
the topic taxonomy.