first_published_at can't be set in the future
There is an issue where if the `first_published_at` is set to any date after the current, it throws and error when trying to edit the edition again.

The `edition.goverment` relies on the `first_published_at` date to associate the edition with the right government and therefore should not be possible to set in the future.