Automatically add legacy associations when tagging
When new content is tagged to the Topic Taxonomy, add some legacy
associations based on the taxons in the Topic Taxonomy that have been
chosen.

This is an interim way of maintaining tagging to these legacy
associations, while not requiring people to do it manually. It should
be removed once there are no more legacy associations to tag to.