Get world taxons separately in EditionTaxonsFetcher
This class gets the taxons that are currently tagged to the
document in the publishing-api. As the publishing-api treats
all taxons as the same, we need to separate world taxons
from topic taxons again here so that we can use
EditionTaxonFetcher to pre-populate the "Selected taxons" box.

Update Statistics Announcements and Publications controller
tests

As EditionsTaxonFetcher is now checking whether taxons are part
of the world taxonomy we need to stub the world taxonomy in these
tests.