Copy all content tagged to Brexit Policy to Brexit Taxon
Every 30 minutes, a rake task will run which checks to see if there are
new content items tagged to the Brexit Policy taxon which are not tagged to
the Brexit Topic Taxonomy. If so, these content items will be tagged to
the new Brexit Topic Taxonomy, this will keep the new taxonomy up to
date with the latest information regarding Brexit.