Modify and organise Topic Taxonomy tagging page
Make all the taxon's that can be tagged to appear in one ordered list on
the Topic Taxonomy tagging interface. This can be done by merging the
alpha, beta and live lists together and ordering all the taxon's in
an alphabetical order.

This will ensure that the Topic tagging interface is easier to understand and
use.