Tweak the Topic Taxonomy tagging message
I don't think this needs to be a warning, or so alarming, as Topic
Taxonomy tagging is done at the document level as the tagging should
be appropriate for any edition.

Therefore, use blue rather than red rather than colour, and update the
text to try to be more informative.