Show the World taxon selection box on the edition review page
Show the world taxons currently tagged to the edition, and links to
a form to edit the worldwide tagging.

1. Get currently tagged world taxons in editions controller

This will be used to populate the select box on the edition
review page that displays which world taxons have been tagged.

2. Create a view to show the selected world topics

Also update publication controller tests to check for presence of World
taxonomy selector.