Create a new list of orgs that can tag to worldwide
The eventual goal is open up tagging to the new topic taxonomy
to all departments, but still keep some restriction on who can
tag to worldwide.

This first step is to keep a completely separate list of who can
tag to worldwide.

Read the list of organisations that can tag to worldwide from
the new list in the tagging configuration set-up.

Update stubbing in tests to match the new configuration.