Check whether to show worldwide tagging interface
Create a new function whether to show the worldwide tagging
interface. From now on, only organisations listed in
`config/worldwide_tagging_organisations.yml` will be able to
tag to world.

Update tests in edition_world_tags_controller_test:

Set the create an organisation with a content_id that can tag
to worldwide.

As all organisations that can tag to the worldwide taxonomy
can also tag to the topic taxonomy we can remove the extra
set-up in the last test.

Fix publications_controller_test:

1. Change the content id of the organisation used in the tests
to one that can tag to both the topic and world taxonomies. This
also us to check for the presence of both the world and topic tagging
interfaces in the view tests. The content id used is DExEU.

2. Update the variable names to reflect that we are using an
organisation that can tag to the worldwide taxonomy.

3. Set the publication type when creating the publication as only
certain document types can tag to the worldwide taxonomy.